face-rec-docker/requirements.txt
boto3
opencv-python-headless
facenet-pytorch
torch
torchvision
numpy
Pillow
